requirements.txt
accelerate==1.6.0
Flask==3.0.3
huggingface-hub==0.30.2
matplotlib==3.10.1
numpy==2.2.5
open3d==0.19.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
pillow==10.4.0
plotly==6.0.1
prompt_toolkit==3.0.51
pyvista==0.45.0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
trimesh==4.6.10
urllib3==2.4.0
